new way to start with a new strategy
