+++ examples/websocket.go
@@ -13,40 +13,40 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-package strategies
+package examples
 
 import (
-	"fmt"
-	"time"
-
 	"github.com/saniales/golang-crypto-trading-bot/environment"
 	"github.com/saniales/golang-crypto-trading-bot/exchanges"
 	"github.com/saniales/golang-crypto-trading-bot/strategies"
+	"github.com/sirupsen/logrus"
 )
 
-// Watch5Min prints out the info of the market every 5 minutes.
-var Watch5Min = strategies.IntervalStrategy{
-	Model: strategies.StrategyModel{
-		Name: "Watch5Min",
-		Setup: func(wrappers []exchanges.ExchangeWrapper, markets []*environment.Market) error {
-			fmt.Println("Watch5Min starting")
-			return nil
+// Websocket strategy
+func Websocket() strategies.Strategy {
+	return strategies.WebsocketStrategy{
+		Model: strategies.StrategyModel{
+			Name: "Websocket",
+			Setup: func(wrappers []exchanges.ExchangeWrapper, markets []*environment.Market) error {
+				for _, wrapper := range wrappers {
+					err := wrapper.FeedConnect(markets)
+					if err == exchanges.ErrWebsocketNotSupported || err == nil {
+						continue
+					}
+					return err
+				}
+				return nil
+			},
+			OnUpdate: func(wrappers []exchanges.ExchangeWrapper, markets []*environment.Market) error {
+				// do something
+				return nil
+			},
+			TearDown: func(wrappers []exchanges.ExchangeWrapper, markets []*environment.Market) error {
+				return nil
+			},
+			OnError: func(err error) {
+				logrus.Error(err)
+			},
 		},
-		OnUpdate: func(wrappers []exchanges.ExchangeWrapper, markets []*environment.Market) error {
-			_, err := wrappers[0].GetMarketSummary(markets[0])
-			if err != nil {
-				return err
-			}
-			fmt.Println(markets)
-			return nil
-		},
-		OnError: func(err error) {
-			fmt.Println(err)
-		},
-		TearDown: func(wrappers []exchanges.ExchangeWrapper, markets []*environment.Market) error {
-			fmt.Println("Watch5Min exited")
-			return nil
-		},
-	},
-	Interval: time.Minute * 5,
+	}
 }
